Key and Operator in Requirements add a lot to the size of InstanceType representation, but I think they're required for reconstruction later. Added an empty test file based on Provisioning which I'll start populating soon.
