+++ pkg/controllers/orb/controller.go
@@ -130,21 +130,21 @@ func (c *Controller) logSchedulingBaselineToPV(item *SchedulingInput) error {
 	fileName := fmt.Sprintf("SchedulingInputBaseline_%s.log", timestampStr)
 	path := filepath.Join("/data", fileName)
 
-	// DEBUG Remove Later
-	fileNametest := fmt.Sprintf("SchedulingInputBaselineTEST_%s.log", timestampStr)
-	pathtest := filepath.Join("/data", fileNametest)
-	file, err := os.Create(pathtest)
-	if err != nil {
-		fmt.Println("Error opening file:", err)
-		return err
-	}
-	defer file.Close()
+	// // DEBUG Remove Later
+	// fileNametest := fmt.Sprintf("SchedulingInputBaselineTEST_%s.log", timestampStr)
+	// pathtest := filepath.Join("/data", fileNametest)
+	// file, err := os.Create(pathtest)
+	// if err != nil {
+	// 	fmt.Println("Error opening file:", err)
+	// 	return err
+	// }
+	// defer file.Close()
 
-	_, err = file.WriteString(item.String())
-	if err != nil {
-		fmt.Println("Error writing data to file:", err)
-		return err
-	}
+	// _, err = file.WriteString(item.String())
+	// if err != nil {
+	// 	fmt.Println("Error writing data to file:", err)
+	// 	return err
+	// }
 
 	fmt.Println("Writing baseline data to S3 bucket.") // test print / remove later
 	return c.writeToPV(logdata, path)
